feat: validate config file (#198)
This PR addresses an issue raised in #190: the need to have a way to validate a configuration file.
The way to validate a configuration file is `kunai config --validate /path/to/config.yaml`.
This initializes tries to initialize kunai with the configuration file and then quits with appropriate error code (0 on success and 1 on error).